+++ src/lib.rs
@@ -27,13 +27,14 @@ extern crate alloc;
 mod decoding;
 mod utils;
 
-/// This module contains functions for decoding extrinsics.
-///
-/// - See [`decode_extrinsic`] for a general function to decode modern or historic extrinsics.
-/// - See [`decode_extrinsic_current`] for a helper to decode modern extrinsics.
-/// - See [`decode_extrinsic_legacy`] for a helper (and example) on decoding legacy extrinsics.
-///
 pub mod extrinsics {
+    //! This module contains functions for decoding extrinsics.
+    //!
+    //! - See [`decode_extrinsic`] for a general function to decode modern or historic extrinsics.
+    //! - See [`decode_extrinsic_current`] for a helper to decode modern extrinsics.
+    //! - See [`decode_extrinsic_legacy`] for a helper (and example) on decoding legacy extrinsics.
+    //!
+
     use crate::utils::InfoAndResolver;
     use scale_type_resolver::TypeResolver;
 
@@ -148,16 +149,17 @@ pub mod extrinsics {
     }
 }
 
-/// This module contains functions for decoding storage keys and values.
-///
-/// - See [`decode_storage_key`] and [`decode_storage_value`] to decode storage keys or values
-///   from modern or historic runtimes.
-/// - See [`decode_storage_key_current`] and [`decode_storage_value_current`] to decode modern
-///   storage keys and values.
-/// - See [`decode_storage_key_legacy`] and [`decode_storage_value_legacy`] to decode historic
-///   storage keys and values (with examples).
-///
 pub mod storage {
+    //! This module contains functions for decoding storage keys and values.
+    //!
+    //! - See [`decode_storage_key`] and [`decode_storage_value`] to decode storage keys or values
+    //!   from modern or historic runtimes.
+    //! - See [`decode_storage_key_current`] and [`decode_storage_value_current`] to decode modern
+    //!   storage keys and values.
+    //! - See [`decode_storage_key_legacy`] and [`decode_storage_value_legacy`] to decode historic
+    //!   storage keys and values (with examples).
+    //!
+
     use crate::utils::InfoAndResolver;
     use scale_decode::Visitor;
     use scale_type_resolver::TypeResolver;
@@ -432,17 +434,18 @@ pub mod storage {
     }
 }
 
-/// Helper functions and types to assist with decoding.
-///
-/// - [`type_registry_from_metadata`] is expected to be used when decoding things from historic
-///   runtimes, adding the ability to decode some types from information in the metadata.
-/// - [`decode_with_error_tracing`] is like [`decode_with_visitor`], but
-///   will use a tracing visitor (if the `error-tracing` feature is enabled) to provide more
-///   information in the event that decoding fails.
-/// - [`list_storage_entries`] returns an iterator over all of the storage entries available in
-///   some metadata.
-///
 pub mod helpers {
+    //! Helper functions and types to assist with decoding.
+    //!
+    //! - [`type_registry_from_metadata`] is expected to be used when decoding things from historic
+    //!   runtimes, adding the ability to decode some types from information in the metadata.
+    //! - [`decode_with_error_tracing`] is like [`decode_with_visitor`], but
+    //!   will use a tracing visitor (if the `error-tracing` feature is enabled) to provide more
+    //!   information in the event that decoding fails.
+    //! - [`list_storage_entries`] returns an iterator over all of the storage entries available in
+    //!   some metadata.
+    //!
+
     pub use crate::utils::{decode_with_error_tracing, DecodeErrorTrace};
     pub use crate::utils::{list_storage_entries, list_storage_entries_any, StorageEntry};
     #[cfg(feature = "legacy")]