+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog].
 
 [Keep a Changelog]: http://keepachangelog.com/en/1.0.0/
 
+## 0.3.0 (2024-09-30)
+
+- Fix `extrinsic.call_range()` and `extensions.range()` functions, and clarify descriptions. 
+- Add `extrinsic.call_args_range()` to return the call data arguments.
+
 ## 0.2.0 (2024-09-30)
 
 - Consistify helper functions; have `list_storage_entries{_any}` and `type_registry_from_metadata{_any}`, where